pyproject.toml
[project]
name = "doc-distill"
version = "0.1.0"
description = "A robust PDF to Markdown converter with LLM prompt compression capabilities."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "pymupdf>=1.23.0",      # For PDF reading
    "pytesseract>=0.3.10",  # For OCR
    "pillow>=10.0.0",       # Image handling for OCR
    "nltk>=3.8.1",          # Natural Language Tool Kit for stop words
    "typer>=0.9.0",         # CLI Framework
    "rich>=13.0.0",         # Beautiful terminal output
]

[project.scripts]
doc-distill = "doc_distill.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc_distill"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
]
